admin 管理页面/requirements.txt
Flask

# Application server for both development and production.
gunicorn
pytest==4.3.0
pytest-cov
flake8
flask-debugtoolbar
Flask-WTF
Flask-SQLAlchemy
Flask-Login
WTForms-Components
psycopg2
Click
sqlalchemy-utils
bcrypt
sendgrid
redis
celery
python-dotenv
flask-moment
Faker
